requirement.txt
flask
python-dotenv
openai
groq
chromadb
azure-cognitiveservices-speech
